Added controllers, updated service and repository
Added controllers routes may have to be changed. Updated services and repos id from int to long, to better follow the models

+++ backend/src/main/java/com/techsage/banking/services/BankAccountServiceJpa.java
@@ -28,9 +28,7 @@ public class BankAccountServiceJpa implements BankAccountService {
     }
 
     @Override
-    public BankAccount getById(int id) {
-        return bankAccountRepository.findById(id).get();
-    }
+    public BankAccount getById(long id) {return bankAccountRepository.findById(id).get();}
 
     @Override
     public BankAccount create(BankAccount bankAccount) {
@@ -43,7 +41,7 @@ public class BankAccountServiceJpa implements BankAccountService {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(long id) {
         bankAccountRepository.deleteById(id);
     }
 }
